feat(commands): add bang option to Config and Plugin
Calling either commands with a bang will create the file if it not
exists. In this case, the picker from telescope will not open. The
extension `.lua` will be appended if not present. You can nest the file.

`:Config! lsp/keymaps` and `:Config! lsp/keymaps.lua` will edit the
file in `lua/user/lsp/keymaps.lua`.

The file is not created until you save it.

`:Config!`, with no args, will open $MYVIMRC.