feat(rooter): Add support to open folder
Sometimes I do `nvim /path/to/folder` so I can browse the folder with
Neotree. At first, I had put this config in the Neotree plugin, but I
think it makes more sense to just support this in rooter itself.

If you open a folder instead of a file, it will set the root to that
folder as if it had autodetected it. This is usually the behavior I
want.